refactor: rename orders check helpers to parcels
